:recycle: Refactor loading pre-fill plugins
These no longer need to load before all the rest because of the form-global
loader that blocks any actions until all module plugins are loaded.
